economic-library/requirements.txt
# Python 依赖包
# 安装：pip install -r requirements.txt

# HTTP 请求
requests>=2.28.0

# 环境变量管理
python-dotenv>=1.0.0

# 数据处理 (可选，用于高级分析)
# pandas>=2.0.0

# Tushare 官方 SDK (可选，当前使用原生 API)
# tushare>=1.2.89
